[FEAT]: Adiciona métodos faltantes dos CRUDS. [ADD]: Adiciona relacionamento de ItemsPedidos. [FIX]: Ajusta retornos da API e tipos de dados

+++ Vendas/src/main/java/com/klayvert/vendas/rest/dtos/ProdutoDTO.java
@@ -6,6 +6,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -15,7 +17,7 @@ public class ProdutoDTO {
 
     private String descricao;
 
-    private float precoUnitario;
+    private BigDecimal precoUnitario;
 
     public static ProdutoDTO toDto(Produto produto){
         ProdutoDTO dto = new ProdutoDTO();
